Refactor Google Maps integration to its own sub component on SearchMap
