+++ src/components/SearchMap/SearchMap.js
@@ -1,80 +1,25 @@
 import React, { Component } from 'react';
 import { arrayOf, bool, func, number, string, shape, object } from 'prop-types';
 import { withRouter } from 'react-router-dom';
-import { withGoogleMap, GoogleMap } from 'react-google-maps';
 import classNames from 'classnames';
-import groupBy from 'lodash/groupBy';
 import isEqual from 'lodash/isEqual';
-import reduce from 'lodash/reduce';
 import routeConfiguration from '../../routeConfiguration';
 import { createResourceLocatorString } from '../../util/routes';
 import { createSlug } from '../../util/urlHelpers';
-import { types as sdkTypes } from '../../util/sdkLoader';
 import { propTypes } from '../../util/types';
 import { obfuscatedCoordinates } from '../../util/maps';
-import { googleBoundsToSDKBounds } from '../../util/googleMaps';
-import { SearchMapInfoCard, SearchMapPriceLabel, SearchMapGroupLabel } from '../../components';
 import config from '../../config';
+import { hasParentWithClassName } from './SearchMap.helpers.js';
 
+import SearchMapWithGoogleMap, {
+  LABEL_HANDLE,
+  INFO_CARD_HANDLE,
+  fitMapToBounds,
+  mapBoundsToSDKBounds,
+  isMapsLibLoaded,
+} from './SearchMapWithGoogleMap';
 import ReusableMapContainer from './ReusableMapContainer';
 import css from './SearchMap.css';
-
-const LABEL_HANDLE = 'SearchMapLabel';
-const INFO_CARD_HANDLE = 'SearchMapInfoCard';
-
-/**
- * Fit part of map (descriped with bounds) to visible map-viewport
- *
- * @param {Object} map - map that needs to be centered with given bounds
- * @param {SDK.LatLngBounds} bounds - the area that needs to be visible when map loads.
- */
-const fitMapToBounds = (map, bounds, padding) => {
-  const { ne, sw } = bounds || {};
-  // map bounds as string literal for google.maps
-  const mapBounds = bounds ? { north: ne.lat, east: ne.lng, south: sw.lat, west: sw.lng } : null;
-
-  // If bounds are given, use it (defaults to center & zoom).
-  if (map && mapBounds) {
-    if (padding == null) {
-      map.fitBounds(mapBounds);
-    } else {
-      map.fitBounds(mapBounds, padding);
-    }
-  }
-};
-
-/**
- * hasParentWithClassName searches class name from parent elements of given target
- * @param {Node} target - element whose parent might contain given class.
- * @param {String} className - class name string to be found
- */
-const hasParentWithClassName = (target, className) => {
-  return [...document.querySelectorAll(`.${className}`)].some(
-    el => el !== target && el.contains(target)
-  );
-};
-
-/**
- * Listings array grouped by geolocation
- * @param {Array} mapListings - listings to be grouped on map
- * @return {Object} - Object where coordinate pair is the key to different listings
- */
-const groupedByCoordinates = mapListings => {
-  return groupBy(mapListings, l => {
-    const g = l.attributes.geolocation;
-    return `${g.lat}-${g.lng}`;
-  });
-};
-
-/**
- * Listings (in location based object literal) is mapped to array
- * @param {Object} mapListings - listings to be grouped on map
- * @return {Array} - An array where items are arrays of listings
- *   (They are arrays containing all the listings in that location)
- */
-const reducedToArray = mapListings => {
-  return reduce(mapListings, (acc, listing) => acc.concat([listing]), []);
-};
 
 const withCoordinatesObfuscated = listings => {
   return listings.map(listing => {
@@ -93,114 +38,12 @@ const withCoordinatesObfuscated = listings => {
   });
 };
 
-/**
- * MapWithGoogleMap uses withGoogleMap HOC.
- * It handles some of the google map initialization states.
- */
-const MapWithGoogleMap = withGoogleMap(props => {
-  const {
-    activeListingId,
-    center,
-    infoCardOpen,
-    listings,
-    onIdle,
-    onListingClicked,
-    onListingInfoCardClicked,
-    createURLToListing,
-    onMapLoad,
-    zoom,
-    mapComponentRefreshToken,
-  } = props;
-
-  const listingArraysInLocations = reducedToArray(groupedByCoordinates(listings));
-  const priceLabels = listingArraysInLocations.reverse().map(listingArr => {
-    const isActive = activeListingId
-      ? !!listingArr.find(l => activeListingId.uuid === l.id.uuid)
-      : false;
-
-    // If location contains only one listing, print price label
-    if (listingArr.length === 1) {
-      const listing = listingArr[0];
-      const infoCardOpenIds = Array.isArray(infoCardOpen) ? infoCardOpen.map(l => l.id.uuid) : [];
-
-      // if the listing is open, don't print price label
-      if (infoCardOpen != null && infoCardOpenIds.includes(listing.id.uuid)) {
-        return null;
-      }
-      return (
-        <SearchMapPriceLabel
-          isActive={isActive}
-          key={listing.id.uuid}
-          className={LABEL_HANDLE}
-          listing={listing}
-          onListingClicked={onListingClicked}
-          mapComponentRefreshToken={mapComponentRefreshToken}
-        />
-      );
-    }
-    return (
-      <SearchMapGroupLabel
-        isActive={isActive}
-        key={listingArr[0].id.uuid}
-        className={LABEL_HANDLE}
-        listings={listingArr}
-        onListingClicked={onListingClicked}
-        mapComponentRefreshToken={mapComponentRefreshToken}
-      />
-    );
-  });
-
-  const listingsArray = Array.isArray(infoCardOpen) ? infoCardOpen : [infoCardOpen];
-  const openedCard = infoCardOpen ? (
-    <SearchMapInfoCard
-      key={listingsArray[0].id.uuid}
-      mapComponentRefreshToken={mapComponentRefreshToken}
-      className={INFO_CARD_HANDLE}
-      listings={listingsArray}
-      onListingInfoCardClicked={onListingInfoCardClicked}
-      createURLToListing={createURLToListing}
-    />
-  ) : null;
-
-  const controlPosition =
-    typeof window !== 'undefined' && typeof window.google !== 'undefined'
-      ? window.google.maps.ControlPosition.LEFT_TOP
-      : 5;
-
-  return (
-    <GoogleMap
-      defaultZoom={zoom}
-      defaultCenter={center}
-      options={{
-        // Disable all controls except zoom
-        mapTypeControl: false,
-        scrollwheel: false,
-        fullscreenControl: false,
-        clickableIcons: false,
-        streetViewControl: false,
-
-        // When infoCard is open, we can't differentiate double click on top of card vs map.
-        disableDoubleClickZoom: !!infoCardOpen,
-
-        zoomControlOptions: {
-          position: controlPosition,
-        },
-      }}
-      ref={onMapLoad}
-      onIdle={onIdle}
-    >
-      {priceLabels}
-      {openedCard}
-    </GoogleMap>
-  );
-});
-
 export class SearchMapComponent extends Component {
   constructor(props) {
     super(props);
 
     this.listings = [];
-    this.googleMap = null;
+    this.mapRef = null;
     this.mapReattachmentCount = 0;
     this.state = { infoCardOpen: null };
 
@@ -212,14 +55,14 @@ export class SearchMapComponent extends Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    if (this.googleMap) {
-      const currentBounds = googleBoundsToSDKBounds(this.googleMap.getBounds());
+    if (this.mapRef) {
+      const currentBounds = mapBoundsToSDKBounds(this.mapRef);
 
       // Do not call fitMapToBounds if bounds are the same.
       // Our bounds are viewport bounds, and fitBounds will try to add margins around those bounds
       // that would result to zoom-loop (bound change -> fitmap -> bounds change -> ...)
       if (!isEqual(nextProps.bounds, currentBounds) && nextProps.useLocationSearchBounds) {
-        fitMapToBounds(this.googleMap, nextProps.bounds, 0);
+        fitMapToBounds(this.mapRef, nextProps.bounds, 0);
       }
     }
   }
@@ -262,11 +105,11 @@ export class SearchMapComponent extends Component {
   }
 
   onMapLoadHandler(map) {
-    this.googleMap = map;
+    this.mapRef = map;
 
-    if (this.googleMap) {
+    if (this.mapRef) {
       // map is ready, let's fit search area's bounds to map's viewport
-      fitMapToBounds(this.googleMap, this.props.bounds, 0);
+      fitMapToBounds(this.mapRef, this.props.bounds, 0);
     }
   }
 
@@ -278,14 +121,12 @@ export class SearchMapComponent extends Component {
       center,
       listings: originalListings,
       mapRootClassName,
-      onCloseAsModal,
       onIdle,
       zoom,
       mapsConfig,
       activeListingId,
     } = this.props;
     const classes = classNames(rootClassName || css.root, className);
-    const mapClasses = mapRootClassName || css.mapRoot;
 
     const listingsWithLocation = originalListings.filter(l => !!l.attributes.geolocation);
     const listings = mapsConfig.fuzzy.enabled
@@ -293,39 +134,33 @@ export class SearchMapComponent extends Component {
       : listingsWithLocation;
     const infoCardOpen = this.state.infoCardOpen;
 
-    const isMapsLibLoaded = typeof window !== 'undefined' && window.google && window.google.maps;
-
     const forceUpdateHandler = () => {
       this.mapReattachmentCount += 1;
     };
 
     // container element listens clicks so that opened SearchMapInfoCard can be closed
     /* eslint-disable jsx-a11y/no-static-element-interactions */
-    return isMapsLibLoaded ? (
+    return isMapsLibLoaded() ? (
       <ReusableMapContainer className={reusableContainerClassName} onReattach={forceUpdateHandler}>
-        <MapWithGoogleMap
+        <SearchMapWithGoogleMap
           containerElement={<div className={classes} onClick={this.onMapClicked} />}
-          mapElement={<div className={mapClasses} />}
+          mapElement={<div className={mapRootClassName || css.mapRoot} />}
           center={center}
+          isOpenOnModal={isOpenOnModal}
+          infoCardOpen={infoCardOpen}
           listings={listings}
           activeListingId={activeListingId}
-          infoCardOpen={infoCardOpen}
+          mapComponentRefreshToken={this.state.mapReattachmentCount}
+          createURLToListing={this.createURLToListing}
           onListingClicked={this.onListingClicked}
           onListingInfoCardClicked={this.onListingInfoCardClicked}
-          createURLToListing={this.createURLToListing}
           onMapLoad={this.onMapLoadHandler}
           onIdle={() => {
-            if (this.googleMap) {
-              onIdle(this.googleMap);
-            }
-          }}
-          onCloseAsModal={() => {
-            if (onCloseAsModal) {
-              onCloseAsModal();
+            if (this.mapRef) {
+              onIdle(this.mapRef);
             }
           }}
           zoom={zoom}
-          mapComponentRefreshToken={this.mapReattachmentCount}
         />
       </ReusableMapContainer>
     ) : (
@@ -341,7 +176,7 @@ SearchMapComponent.defaultProps = {
   mapRootClassName: null,
   reusableContainerClassName: null,
   bounds: null,
-  center: new sdkTypes.LatLng(0, 0),
+  center: null,
   activeListingId: null,
   listings: [],
   onCloseAsModal: null,
