+++ src/components/SearchMapPriceLabel/SearchMapPriceLabel.js
@@ -1,7 +1,5 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { OverlayView } from 'react-google-maps';
-import { OVERLAY_VIEW } from 'react-google-maps/lib/constants';
 import { injectIntl, intlShape } from 'react-intl';
 import classNames from 'classnames';
 import { propTypes } from '../../util/types';
@@ -10,26 +8,6 @@ import { ensureListing } from '../../util/data';
 import config from '../../config';
 
 import css from './SearchMapPriceLabel.css';
-
-// FIX "TypeError: Cannot read property 'overlayMouseTarget' of null"
-// Override draw function to catch errors with map panes being undefined to prevent console errors
-// https://github.com/tomchentw/react-google-maps/issues/482
-class CustomOverlayView extends OverlayView {
-  draw() {
-    // https://developers.google.com/maps/documentation/javascript/3.exp/reference#MapPanes
-    const mapPanes = this.state[OVERLAY_VIEW].getPanes();
-    // Add conditional to ensure panes and container exist before drawing
-    if (mapPanes && this.containerElement) {
-      super.draw();
-    }
-  }
-}
-
-// Center label so that caret is pointing to correct pixel.
-// (vertical positioning: height + arrow) */
-const getPixelPositionOffset = (width, height) => {
-  return { x: -1 * (width / 2), y: -1 * (height + 3) };
-};
 
 class SearchMapPriceLabel extends Component {
   shouldComponentUpdate(nextProps) {
@@ -45,30 +23,22 @@ class SearchMapPriceLabel extends Component {
   render() {
     const { className, rootClassName, intl, listing, onListingClicked, isActive } = this.props;
     const currentListing = ensureListing(listing);
-    const { geolocation, price } = currentListing.attributes;
+    const { price } = currentListing.attributes;
 
     // Create formatted price if currency is known or alternatively show just the unknown currency.
     const formattedPrice =
       price && price.currency === config.currency ? formatMoney(intl, price) : price.currency;
 
-    // Explicit type change to object literal for Google OverlayViews (geolocation is SDK type)
-    const latLngLiteral = { lat: geolocation.lat, lng: geolocation.lng };
     const classes = classNames(rootClassName || css.root, className);
     const priceLabelClasses = classNames(css.priceLabel, { [css.priceLabelActive]: isActive });
     const caretClasses = classNames(css.caret, { [css.caretActive]: isActive });
 
     return (
-      <CustomOverlayView
-        position={latLngLiteral}
-        mapPaneName={OverlayView.OVERLAY_MOUSE_TARGET}
-        getPixelPositionOffset={getPixelPositionOffset}
-      >
-        <button className={classes} onClick={() => onListingClicked(currentListing)}>
-          <div className={css.caretShadow} />
-          <div className={priceLabelClasses}>{formattedPrice}</div>
-          <div className={caretClasses} />
-        </button>
-      </CustomOverlayView>
+      <button className={classes} onClick={() => onListingClicked(currentListing)}>
+        <div className={css.caretShadow} />
+        <div className={priceLabelClasses}>{formattedPrice}</div>
+        <div className={caretClasses} />
+      </button>
     );
   }
 }
